requirement.txt
langchain
langchain_core
langchain_community
openai
python-dotenv
streamlit
google.genai